feat(issue-634): addressing memory leak
